docs: adiciona comentários explicativos nos exercícios

+++ exercicios/exerc-03.py
@@ -1,8 +1,12 @@
-largura = float(input("Digite a largura do retângulo: "))
-altura = float(input("Digite a altura do retângulo: "))
+# int() converte o valor digitado para número inteiro
+
+largura = int(input("Digite a largura do retângulo: "))
+altura = int(input("Digite a altura do retângulo: "))
 
 area = largura * altura
 perimetro = 2 * (largura + altura)
 
 print(f"Área do retângulo: {area}")
 print(f"Perímetro do retângulo: {perimetro}")
+
+
